Add --allowMissingData to VariantsToTable
In GATK 3.6 "If a VCF record is missing a value, then the tool by
default throws an error, but the special value NA can be emitted instead
if requested at the command line using --allowMissingData."

Depending on the arguments, we might encounter missing values which
would cause an error without this flag.

This behavior was reversed in GATK 4.0 and it will no longer throw an
error on missing values.