add joint sv filtering for depth callers
This runs the depth filters for events called by cnvkit/cnvnator. Final
sample names follow the $SAMPLE-$CALLER format. This allows easy
tracking for the source of calls in final merged vcfs.
added custom merge sv records. This allows calls to be merged together
if they are of the same type and within a bp window. This does not
remove calls just adds a new record in the output vcf.